requirements.txt
flask
pdfplumber
pymupdf
pytesseract
pillow
networkx
matplotlib
pypinyin
networkx